# backend/docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Sports Management API
  description: |
    Comprehensive API for managing sports leagues, referee assignments, games, teams, and budgets.
    
    ## Authentication
    All endpoints (except `/health` and `/auth/login`) require Bearer authentication.
    Include the JWT token in the Authorization header: `Authorization: Bearer <token>`
    
    ## Rate Limiting
    - Login endpoint: 5 requests per minute
    - Registration endpoint: 3 requests per hour
    - Password reset: 3 requests per hour
    - Assignment operations: 100 requests per minute
    - General endpoints: 1000 requests per hour
    
    ## Error Handling
    All error responses follow a consistent format:
    ```json
    {
      "error": "Error message",
      "message": "Detailed description",
      "code": "ERROR_CODE"
    }
    ```
    
    ## Data Models
    The API follows a hierarchical structure:
    - Organizations contain Leagues
    - Leagues contain Teams
    - Games reference Teams and are assigned Referees
    - Budgets track financial aspects
    
  version: 1.0.0
  contact:
    name: Sports Management API Support
    email: support@sportsmanagement.app
  license:
    name: MIT
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.sportsmanagement.app
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  schemas:
...
    Assignment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        game_id:
          type: string
          format: uuid
          description: Reference to game
        user_id:
          type: string
          format: uuid
          description: Reference to referee user
        position_id:
          type: string
          format: uuid
          description: Reference to position
        status:
          type: string
          enum: [pending, accepted, declined, completed]
          description: Assignment status
        assigned_at:
          type: string
          format: date-time
          description: When the assignment was made
        assigned_by:
          type: string
          format: uuid
          description: User who made the assignment
        calculated_wage:
          type: number
          minimum: 0
          description: Final calculated wage for this assignment
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - game_id
        - user_id
        - position_id
...
    CreateAssignmentRequest:
      type: object
      properties:
        game_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        position_id:
          type: string
          format: uuid
        assigned_by:
          type: string
          format: uuid
      required:
        - game_id
        - user_id
        - position_id
    BulkUpdateAssignmentsRequest:
      type: object
      properties:
        updates:
          type: array
          items:
            type: object
            properties:
              assignment_id:
                type: string
                format: uuid
              status:
                type: string
                enum: [pending, accepted, declined, completed]
              calculated_wage:
                type: number
                minimum: 0
            required:
              - assignment_id
              - status
          maxItems: 100
      required:
        - updates
    BulkRemoveAssignmentsRequest:
      type: object
      properties:
        assignment_ids:
          type: array
          items:
            type: string
            format: uuid
          maxItems: 100
      required:
        - assignment_ids
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        message:
          type: string
          description: Detailed error description
        code:
          type: string
          description: Error code for programmatic handling
...
security:
  - bearerAuth: []
paths:
...
  /assignments:
    get:
      tags:
        - Assignments
      summary: Get all assignments
      description: Retrieve all referee assignments with optional filtering
      parameters:
        - name: game_id
          in: query
          description: Filter by game ID
          schema:
            type: string
            format: uuid
        - name: referee_id
          in: query
          description: Filter by referee ID
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          description: Filter by assignment status
          schema:
            type: string
            enum: [pending, accepted, declined, completed]
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Assignments retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'

    post:
      tags:
        - Assignments
      summary: Create new assignment
      description: Create a new referee assignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssignmentRequest'
            example:
              game_id: "123e4567-e89b-12d3-a456-426614174000"
              user_id: "123e4567-e89b-12d3-a456-426614174001"
              position_id: "123e4567-e89b-12d3-a456-426614174002"
              assigned_by: "123e4567-e89b-12d3-a456-426614174003"
      responses:
        '201':
          description: Assignment created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      assignment:
                        $ref: '#/components/schemas/Assignment'
                      wageBreakdown:
                        type: object
                        description: Detailed wage calculation breakdown
                  warnings:
                    type: array
                    items:
                      type: string
                    description: Any warnings about the assignment
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Game, referee, or position not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Assignment conflict (position filled, referee already assigned, etc.)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
                  conflicts:
                    type: array
                    items:
                      type: object
  /assignments/{id}:
    get:
      tags:
        - Assignments
      summary: Get specific assignment
      description: Retrieve a specific assignment by ID
      parameters:
        - name: id
          in: path
          required: true
          description: Assignment ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Assignment retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'
        '404':
          description: Assignment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Assignments
      summary: Remove assignment
      description: Remove a referee assignment
      parameters:
        - name: id
          in: path
          required: true
          description: Assignment ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Assignment removed successfully
        '404':
          description: Assignment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /assignments/{id}/status:
    patch:
      tags:
        - Assignments
      summary: Update assignment status
      description: Update the status of an assignment
      parameters:
        - name: id
          in: path
          required: true
          description: Assignment ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [pending, accepted, declined, completed]
              required:
                - status
            example:
              status: "accepted"
      responses:
        '200':
          description: Assignment status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      assignment:
                        $ref: '#/components/schemas/Assignment'
        '404':
          description: Assignment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /assignments/bulk-update:
    post:
      tags:
        - Assignments
      summary: Bulk update assignments
      description: Update multiple assignment statuses at once (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkUpdateAssignmentsRequest'
            example:
              updates:
                - assignment_id: "123e4567-e89b-12d3-a456-426614174000"
                  status: "accepted"
                - assignment_id: "123e4567-e89b-12d3-a456-426614174001"
                  status: "declined"
      responses:
        '200':
          description: Bulk update completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      updatedAssignments:
                        type: array
                        items:
                          $ref: '#/components/schemas/Assignment'
                      summary:
                        type: object
                        properties:
                          totalRequested:
                            type: integer
                          successfulUpdates:
                            type: integer
                          failedUpdates:
                            type: integer
                  warnings:
                    type: array
                    items:
                      type: string
                  partialSuccess:
                    type: boolean
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /assignments/bulk-remove:
    delete:
      tags:
        - Assignments
      summary: Bulk remove assignments
      description: Remove multiple assignments at once (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkRemoveAssignmentsRequest'
            example:
              assignment_ids:
                - "123e4567-e89b-12d3-a456-426614174000"
                - "123e4567-e89b-12d3-a456-426614174001"
      responses:
        '200':
          description: Bulk removal completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      deletedCount:
                        type: integer
                      affectedGames:
                        type: array
                        items:
                          type: string
                          format: uuid
                      summary:
                        type: object
                        properties:
                          requested:
                            type: integer
                          successfullyDeleted:
                            type: integer
                          notFound:
                            type: integer
                  warnings:
                    type: array
                    items:
                      type: object
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /assignments/check-conflicts:
    post:
      tags:
        - Assignments
      summary: Check assignment conflicts
      description: Check for conflicts before creating an assignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssignmentRequest'
      responses:
        '200':
          description: Conflict check completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      hasConflicts:
                        type: boolean
                      conflicts:
                        type: array
                        items:
                          type: string
                      warnings:
                        type: array
                        items:
                          type: string
                      errors:
                        type: array
                        items:
                          type: string
                      isQualified:
                        type: boolean
                      canAssign:
                        type: boolean
  /assignments/available-referees/{game_id}:
    get:
      tags:
        - Assignments
      summary: Get available referees for game
      description: Get list of referees available for a specific game
      parameters:
        - name: game_id
          in: path
          required: true
          description: Game ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Available referees retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      referees:
                        type: array
                        items:
                          allOf:
                            - $ref: '#/components/schemas/User'
                            - type: object
                              properties:
                                availabilityScore:
                                  type: number
                                availabilityStatus:
                                  type: string
                                  enum: [available, not_specified, unknown, conflict]
                                availabilityNote:
                                  type: string
                                availabilityWindows:
                                  type: array
                                  items:
                                    type: object
                      gameTime:
                        type: object
                        properties:
                          date:
                            type: string
                            format: date
                          startTime:
                            type: string
                          endTime:
                            type: string
                      summary:
                        type: object
                        properties:
                          total:
                            type: integer
                          available:
                            type: integer
                          notSpecified:
                            type: integer
                          unknown:
                            type: integer
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # AI Suggestions
...